third_party/WebKit/LayoutTests/fast/events/click-focus-anchor-expected.txt

This test ensures that we can click to focus an a element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


a1 received focus (focusable)
a4 received focus (focusable)
a5 received focus (focusable)
p6 received focus (focusable)
p7 received focus (focusable)
PASS successfullyParsed is true

TEST COMPLETE
